+++ packages/integration-sdk/CHANGELOG.md
@@ -9,6 +9,19 @@ and this project adheres to
 
 ## Unreleased
 
+### Added
+
+- `ExecutionHistory.current: Execution` to provide information about the current
+  execution.
+
+### Changed
+
+- `ExecutionHistory` properties have been renamed as part of adding
+  `current: Execution`, to avoid duplication in the naming: `lastExecution` ->
+  `previous`, `lastSuccessfulExecution` -> `lastSuccessful`.
+- BREAKING: `executeIntegrationInstance` now requires an `ExecutionHistory`
+  argument containing the `current: Execution`.
+
 ## 4.3.0 - 2020-11-24
 
 ### Added